@@ -13,11 +13,11 @@
     ],
     "require": {
         "ext-pcre": "*",
-        "php": ">=7.2.0",
+        "php": ">=8.1",
         "jakeasmith/http_build_url": "^1"
     },
     "require-dev": {
-        "phpunit/phpunit": "8.0"
+        "phpunit/phpunit": "^9 || ^10"
     },
     "suggest":  {
         "ext-mbstring": "Install ext/mbstring for using input / output other than UTF-8 or ISO-8859-1",
