@@ -1,9 +1,10 @@
 <?php
-namespace Algo26\IdnaConvert\Test;
+namespace Algo26\IdnaConvert\Test\integration;
 
 use Algo26\IdnaConvert\Exception\AlreadyPunycodeException;
 use Algo26\IdnaConvert\Exception\InvalidCharacterException;
 use Algo26\IdnaConvert\Exception\InvalidIdnVersionException;
+use Algo26\IdnaConvert\Exception\Std3AsciiRulesViolationException;
 use Algo26\IdnaConvert\ToIdn;
 use PHPUnit\Framework\TestCase;
 
@@ -19,7 +20,7 @@ class ToIdnTest extends TestCase
      * @throws InvalidCharacterException
      * @throws InvalidIdnVersionException
      */
-    public function testEncodeUtf8($decoded, $expectEncoded)
+    public function testEncodeUtf8($decoded, $expectEncoded): void
     {
         $idnaConvert = new ToIdn();
         $encoded = $idnaConvert->convert($decoded);
@@ -42,7 +43,7 @@ class ToIdnTest extends TestCase
      * @throws InvalidCharacterException
      * @throws InvalidIdnVersionException
      */
-    public function testEncodeUtf8Idna2003($decoded, $expectEncoded)
+    public function testEncodeUtf8Idna2003($decoded, $expectEncoded): void
     {
         $idnaConvert = new ToIdn(2003);
         $encoded = $idnaConvert->convert($decoded);
@@ -63,7 +64,7 @@ class ToIdnTest extends TestCase
      *
      * @throws InvalidIdnVersionException
      */
-    public function testEncodeEmailAddress($decoded, $expectEncoded)
+    public function testEncodeEmailAddress($decoded, $expectEncoded): void
     {
         $idnaConvert = new ToIdn(2008);
         $encoded = $idnaConvert->convertEmailAddress($decoded);
@@ -84,7 +85,7 @@ class ToIdnTest extends TestCase
      *
      * @throws InvalidIdnVersionException
      */
-    public function testEncodeUrl($decoded, $expectEncoded)
+    public function testEncodeUrl($decoded, $expectEncoded): void
     {
         $idnaConvert = new ToIdn(2008);
         $encoded = $idnaConvert->convertUrl($decoded);
@@ -100,7 +101,40 @@ class ToIdnTest extends TestCase
         );
     }
 
-    public function providerUtf8()
+    /**
+     * @dataProvider providerAlreadyPunycode
+     */
+    public function testThrowsAlreadyPunycodeException($decoded, $idnVersion): void
+    {
+        self::expectException(AlreadyPunycodeException::class);
+
+        $idnaConvert = new ToIdn($idnVersion);
+        $idnaConvert->convert($decoded);
+    }
+
+    /**
+     * @dataProvider providerInvalidCharacter
+     */
+    public function testThrowsInvalidCharacterException($sequence): void
+    {
+        self::expectException(InvalidCharacterException::class);
+
+        $idnaConvert = new ToIdn(2008);
+        $idnaConvert->convert($sequence);
+    }
+
+    /**
+     * @dataProvider providerStd3AsciiRulesViolation
+     */
+    public function testThrowsStd3AsciiRulesException($sequence): void
+    {
+        self::expectException(Std3AsciiRulesViolationException::class);
+
+        $idnaConvert = new ToIdn(2008, true);
+        $idnaConvert->convert($sequence);
+    }
+
+    public function providerUtf8(): array
     {
         return [
             ['', ''],
@@ -114,7 +148,6 @@ class ToIdnTest extends TestCase
             ['æšŧüø.example', 'xn--6ca0bl71b4a.example'],
             ['ìåíèäæìúíò.example', 'xn--4cabegsede9b0e.example'],
             ['мениджмънт.example', 'xn--d1abegsede9b0e.example'],
-            ['3+1', '3+1'],
             ['www.bäckermüller.example', 'www.xn--bckermller-q5a70a.example'],
             ['ı', 'xn--cfa'],
             ['ekşisözlük', 'xn--ekiszlk-d1a0dy4d'],
@@ -148,17 +181,61 @@ class ToIdnTest extends TestCase
             ['déjà.vu.example', 'xn--dj-kia8a.vu.example'],
             ['efraín.example', 'xn--efran-2sa.example'],
             ['ñandú.example', 'xn--and-6ma2c.example'],
-            ['Foo.âBcdéf.example', 'Foo.xn--bcdf-9na9b.example'],
+            ['Foo.âBcdéf.example', 'foo.xn--bcdf-9na9b.example'],
             ['موقع.وزارة-الاتصالات.مصر', 'xn--4gbrim.xn----ymcbaaajlc6dj7bxne2c.xn--wgbh1c'],
             ['fußball.example', 'xn--fuball-cta.example'],
             ['היפא18פאטאם', 'xn--18-uldcat6ad6bydd'],
             ['فرس18النهر', 'xn--18-dtd1bdi0h3ask'],
             ["\u{33c7}", 'xn--czk'],
             ["\u{37a}", 'xn--1va'],
+            ['ídn', 'xn--dn-mja'],
+            ['ëx.ídn', 'xn--x-ega.xn--dn-mja'],
+            ['åþç', 'xn--5cae2e'],
+            ['ăbĉ', 'xn--b-rhat'],
+            ['ȧƀƈ', 'xn--lhaq98b'],
+            ['ḁḃḉ', 'xn--2fges'],
+            ['丿人尸', 'xn--xiqplj17a'],
+            ['かがき', 'xn--u8jcd'],
+            ['カガキ', 'xn--lckcd'],
+            ['각', 'xn--p39a'],
+            ['걩듆쀺', 'xn--o69aq2nl0j'],
+            ['ꀊꀠꊸ', 'xn--6l7arby7j'],
+            ['αβγ', 'xn--mxacd'],
+            ['ἂἦὕ', 'xn--fng7dpg'],
+            ['абв', 'xn--80acd'],
+            ['աբգ', 'xn--y9acd'],
+            ['აბგ', 'xn--lodcd'],
+            ['∡↺⊂', 'xn--b7gxomk'],
+            ['कखग', 'xn--11bcd'],
+            ['কখগ', 'xn--p5bcd'],
+            ['ਕਖਗ', 'xn--d9bcd'],
+            ['કખગ', 'xn--0dccd'],
+            ['କଖଗ', 'xn--ohccd'],
+            ['கஙச', 'xn--clcid'],
+            ['కఖగ', 'xn--zoccd'],
+            ['ಕಖಗ', 'xn--nsccd'],
+            ['കഖഗ', 'xn--bwccd'],
+            ['කඛග', 'xn--3zccd'],
+            ['กขฃ', 'xn--12ccd'],
+            ['ກຂຄ', 'xn--p6ccg'],
+            ['ཀཁག', 'xn--5cdcd'],
+            ['ကခဂ', 'xn--nidcd'],
+            ['កខគ', 'xn--i2ecd'],
+            ['ᠠᠡᠢ', 'xn--26ecd'],
+            ['ابة', 'xn--mgbcd'],
+            ['אבג', 'xn--4dbcd'],
+            ['ܐܑܒ', 'xn--9mbcd'],
+            ['abcカガキ', 'xn--abc-mj4bfg'],
+            ['åþçカガキ', 'xn--5cae2e328wfag'],
+            ['¹1', '11'],
+            ['Ⅵvi', 'vivi'],
+            ['3002-test。ídn', '3002-test.xn--dn-mja'],
+            ['ff0e-test．ídn', 'ff0e-test.xn--dn-mja'],
+            ['ff61-test｡ídn', 'ff61-test.xn--dn-mja'],
         ];
     }
 
-    public function providerUtf8Idna2003()
+    public function providerUtf8Idna2003(): array
     {
         return [
             ['daß.example', 'dass.example'],
@@ -172,7 +249,7 @@ class ToIdnTest extends TestCase
         ];
     }
 
-    public function providerEmailAddress()
+    public function providerEmailAddress(): array
     {
         return [
             ['some.user@мениджмънт.example', 'some.user@xn--d1abegsede9b0e.example'],
@@ -182,7 +259,7 @@ class ToIdnTest extends TestCase
         ];
     }
 
-    public function providerUrl()
+    public function providerUrl(): array
     {
         return [
             [
@@ -209,6 +286,37 @@ class ToIdnTest extends TestCase
                 'file:///some/path/sömewhere/',
                 'file:///some/path/sömewhere/'
             ],
+        ];
+    }
+
+    public function providerAlreadyPunycode(): array
+    {
+        return [
+            ['xn--ïdn', 2003],
+            ['ⅹn--ädn', 2008],
+            ['xN--ïdn', 2003],
+            ['Xn--ïdn', 2003],
+            ['XN--ïdn', 2003],
+        ];
+    }
+
+    public function providerInvalidCharacter(): array
+    {
+        return [
+            ['3+1'],
+            ['abc+def'],
+            ['do you copy?'],
+            ['yes, minister!'],
+        ];
+    }
+
+    public function providerStd3AsciiRulesViolation(): array
+    {
+        return [
+            ['-hyphenated'],
+            ['-hyphenated-'],
+            ['hyphenated-'],
+            ['negative-test-for-now'],
         ];
     }
 }
